refactor(backup/gcs): inline manifest types
