--- backup-stores/gcs/src/main/java/io/camunda/zeebe/backup/gcs/manifest/CompletedManifest.java
@@ -1,13 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Zeebe Community License 1.1. You may not use this file
- * except in compliance with the Zeebe Community License 1.1.
- */
-package io.camunda.zeebe.backup.gcs.manifest;
-
-public interface CompletedManifest extends Manifest {
-
-  FailedManifest fail(final String failureReason);
-}
